Fix
- Complete overhaul of the random result algorithm, because the previous one was not working correctly.
- Fixed a bug in the "calculateWinProbability" and "calculateNewRating" functions, which previously received a wrong result due to operation between different types (float and integer).
NOTE: Due to the addition of this new random algorithm and its output independence between standard and concurrent version, the correctness test will likely fail multiple times. To avoid many printfs, I decided to leave the test as a comment.  I hope to find a solution for this issue in the near future.